Fixed Chapter 1

+++ Unit-02/Chapter-01/02-01-01.py
@@ -1,1 +1,1 @@
-print(str(int(input()) > 100))
+print(str(float(input()) > 100))
